% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{authenticate_service_account}
\alias{authenticate_service_account}
\title{Authenticate Datastore using a service account}
\usage{
authenticate_service_account(credentials, project)
}
\arguments{
\item{credentials}{Environmental variable or service account.}

\item{project}{Google cloud platform project id/name.}
}
\description{
Set credentials to the name of an environmental variable
or the location of your service account json key file.
}
\seealso{
\url{https://cloud.google.com/}

\url{https://cloud.google.com/datastore/docs/concepts/overview}

\url{https://developers.google.com/identity/protocols/OAuth2#basicsteps}
}

