% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_df.R
\name{commit_df}
\alias{commit_df}
\title{commit_df}
\usage{
commit_df(kind = NULL, name = NULL, df, mutation_type = "upsert",
  keep_existing = TRUE)
}
\arguments{
\item{kind}{dataframe name}

\item{name}{name column to use.}
}
\value{
Data frame of results
}
\description{
Query using the Google Query Language (GQL)
}
\examples{
gql("SELECT * FROM test")

}
\seealso{
\url{https://cloud.google.com/datastore/docs/apis/gql/gql_reference} - The google query language
}

