% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{gql}
\alias{gql}
\title{gql}
\usage{
gql(query)
}
\arguments{
\item{query}{The query to run.}
}
\value{
Data frame of results
}
\description{
Query using the Google Query Language (GQL)
}
\examples{
gql("SELECT * FROM test")

}
\seealso{
\url{https://cloud.google.com/datastore/docs/apis/gql/gql_reference} - The google query language
}

