% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lookup}
\alias{lookup}
\title{Lookup}
\usage{
lookup(kind, name = NULL, id = NULL)
}
\arguments{
\item{kind}{The entity kind}

\item{name}{The entity name.}

\item{id}{The entity id.}

\item{project}{Google cloud platform project id/name.}
}
\value{
dataframe of entity.
}
\description{
Lookup
}
\seealso{
\url{https://cloud.google.com/datastore/docs/concepts/entities} - Entities, Properties, and Keys
}

