% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transaction}
\alias{transaction}
\title{Transaction}
\usage{
transaction()
}
\arguments{
\item{kind}{The entity kind}

\item{name}{The name or id}

\item{...}{Additional properties to store}
}
\description{
Transaction
}
\seealso{
\url{https://cloud.google.com/datastore/docs/concepts/entities} - Entities, Properties, and Keys
}

