% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{commit}
\alias{commit}
\title{commit}
\usage{
commit(kind, name = NULL, ..., mutation_type = "upsert")
}
\arguments{
\item{kind}{The entity kind}

\item{name}{The name of the entity. If NULL, id is autogenerated.}

\item{...}{Additional arguments to store as properties.}

\item{mutation_type}{The type of mutation. One of \strong{insert}, \strong{update}, \strong{upsert}, or \strong{delete}. Default is upsert.}
}
\description{
Commit a single entity to the datastore.
}
\seealso{
\url{https://cloud.google.com/datastore/docs/concepts/entities} - Entities, Properties, and Keys
}

