% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{commit}
\alias{commit}
\title{commit}
\usage{
commit(kind, name = NULL, ..., mutation_type = "upsert",
  keep_existing = TRUE)
}
\arguments{
\item{kind}{The entity kind}

\item{name}{The name of the entity. If NULL, id is autogenerated.}

\item{...}{Additional arguments to store as properties.}

\item{mutation_type}{The type of mutation. One of \strong{insert}, \strong{update}, \strong{upsert}, or \strong{delete}. Default is upsert.}

\item{keep_existing}{Preserve/edit existing data. The entity will be retrieved and specified properties replaced or new properties added. Existing properties will be retained. Default is TRUE.}
}
\value{
list of content and the transaction id
}
\description{
Commit a single entity to the datastore.
Specify a \code{kind} and optionally a \code{name} (setting \code{name = NULL} will result in an id being assigned as a substitute for name).
Additional arguments are treated as properties of the entity. \code{mutation_type} can be set to one of the following:
}
\examples{
commit("test", "entity_name", int_column = as.integer(1), str_column = "awesome!", float_column = 3.14)

}
\seealso{
\url{https://cloud.google.com/datastore/docs/concepts/entities} - Entities, Properties, and Keys
}

